Rename view package to screen, move files
