+++ core/src/main/kotlin/no/ntnu/beardblaster/BeardBlasterGame.kt
@@ -10,9 +10,7 @@ import ktx.app.KtxGame
 import ktx.log.debug
 import ktx.log.logger
 import no.ntnu.beardblaster.assets.Assets
-import no.ntnu.beardblaster.user.UserAuth
-import no.ntnu.beardblaster.utils.AbstractScreen
-import no.ntnu.beardblaster.view.*
+import no.ntnu.beardblaster.screen.*
 
 private val LOG = logger<BeardBlasterGame>()
 const val worldWidth = 1920f
@@ -28,10 +26,7 @@ class BeardBlasterGame : KtxGame<AbstractScreen>() {
         // Set debug level
         Gdx.app.logLevel = Application.LOG_DEBUG
         LOG.debug { "Create game instance" }
-        // Speed login
-        if (!UserAuth().isLoggedIn()) {
-            UserAuth().signIn("beard@blaster.com", "beardblaster")
-        }
+        
         initScreens()
     }
