+++ core/src/main/kotlin/no/ntnu/beardblaster/screen/BeardBlasterStage.kt
@@ -1,4 +1,4 @@
-package no.ntnu.beardblaster.utils
+package no.ntnu.beardblaster.screen
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.scenes.scene2d.Stage
