+++ core/src/main/kotlin/no/ntnu/beardblaster/screen/RegisterScreen.kt
@@ -1,4 +1,4 @@
-package no.ntnu.beardblaster.view
+package no.ntnu.beardblaster.screen
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.Color
@@ -12,67 +12,65 @@ import ktx.log.logger
 import no.ntnu.beardblaster.BeardBlasterGame
 import no.ntnu.beardblaster.assets.Assets
 import no.ntnu.beardblaster.user.UserAuth
-import no.ntnu.beardblaster.utils.AbstractScreen
-import no.ntnu.beardblaster.utils.BeardBlasterStage
 
-private val LOG = logger<LoginScreen>()
 
-class LoginScreen(game: BeardBlasterGame) : AbstractScreen(game) {
+private val LOG = logger<RegisterScreen>()
+
+class RegisterScreen(game: BeardBlasterGame) : AbstractScreen(game) {
     private lateinit var skin: Skin
-    private lateinit var table: Table
-    private lateinit var heading: Label
 
+    private lateinit var table: Table
     private lateinit var leftTable: Table
     private lateinit var rightTable: Table
+    private lateinit var heading: Label
 
-    private lateinit var loginButton: TextButton
-    private lateinit var backButton: Button
+    private lateinit var backBtn: Button
+    private lateinit var createBtn: TextButton
 
+    private lateinit var userNameInput: TextField
     private lateinit var emailInput: TextField
     private lateinit var passwordInput: TextField
+    private lateinit var rePasswordInput: TextField
 
-    private val loginStage: Stage by lazy {
+    private val registrationStage: Stage by lazy {
         val result = BeardBlasterStage()
         Gdx.input.inputProcessor = result
         result
     }
 
     override fun show() {
-        LOG.debug { "LOGIN SCREEN" }
-
-        Gdx.input.inputProcessor = loginStage
+        LOG.debug { "REGISTRATION Screen" }
+        Gdx.input.inputProcessor = registrationStage
         skin = Skin(Assets.assetManager.get(Assets.atlas))
         table = Table(skin)
         table.setBounds(0f, 0f, viewport.worldWidth, viewport.worldHeight)
-
         rightTable = Table(skin)
         leftTable = Table(skin)
-
 
         val standardFont = Assets.assetManager.get(Assets.standardFont)
 
         Label.LabelStyle(standardFont, Color.BLACK).also {
-            heading = Label("Log in", it)
+            heading = Label("Create Wizard", it)
             heading.setFontScale(2f)
             it.background = skin.getDrawable("modal_fancy_header")
             heading.setAlignment(Align.center)
         }
 
-        val buttonStyle = TextButton.TextButtonStyle()
-        skin.getDrawable("button_default_pressed").also { buttonStyle.down = it }
-        skin.getDrawable("button_default").also { buttonStyle.up = it }
+        val createUserButtonStyle = TextButton.TextButtonStyle()
+        skin.getDrawable("button_default_pressed").also { createUserButtonStyle.down = it }
+        skin.getDrawable("button_default").also { createUserButtonStyle.up = it }
 
-        standardFont.apply {
-            buttonStyle.font = this
+        val backBtnStyle = Button.ButtonStyle()
+        skin.getDrawable("modal_fancy_header_button_red_cross_left").also {
+            backBtnStyle.down = it
+            backBtnStyle.up = it
         }
 
-        val backButtonStyle = Button.ButtonStyle()
-        skin.getDrawable("modal_fancy_header_button_red_cross_left").also { backButtonStyle.down = it }
-        skin.getDrawable("modal_fancy_header_button_red_cross_left").also { backButtonStyle.up = it }
-
-
-        loginButton = TextButton("LOGIN", buttonStyle)
-        backButton = Button(backButtonStyle)
+        standardFont.apply {
+            createUserButtonStyle.font = this
+        }
+        backBtn = Button(backBtnStyle)
+        createBtn = TextButton("CREATE WIZARD", createUserButtonStyle)
         setBtnEventListeners()
 
         val textInputStyle = TextField.TextFieldStyle()
@@ -84,30 +82,38 @@ class LoginScreen(game: BeardBlasterGame) : AbstractScreen(game) {
             it.messageFontColor = Color.GRAY
         }
 
+        userNameInput = TextField("", textInputStyle)
+        userNameInput.messageText = "Enter wizard name.."
         emailInput = TextField("", textInputStyle)
         emailInput.messageText = "Enter e-mail address.."
         passwordInput = TextField("", textInputStyle)
         passwordInput.setPasswordCharacter("*"[0])
         passwordInput.isPasswordMode = true
         passwordInput.messageText = "Enter password.."
+        rePasswordInput = TextField("", textInputStyle)
+        rePasswordInput.setPasswordCharacter("*"[0])
+        rePasswordInput.messageText = "Re-enter password.."
+        rePasswordInput.isPasswordMode = true
 
-
-        // TODO: find out why input fields renders with wrong width
-        // Creating table
         rightTable.apply {
             this.defaults().pad(30f)
             this.background = skin.getDrawable("modal_fancy")
             this.add(heading)
             this.row()
+            this.add(userNameInput).width(570f)
+            this.row()
             this.add(emailInput).width(570f)
             this.row()
             this.add(passwordInput).width(570f)
             this.row()
-            this.add(loginButton)
+            this.add(rePasswordInput).width(570f)
+            this.row()
+            this.add(createBtn).width(370f)
         }
 
         val stack = Stack()
-        stack.add(backButton)
+        stack.add(backBtn)
+
         leftTable.apply {
             this.top()
             this.add(stack).fillY().align(Align.top).padTop(50f)
@@ -119,23 +125,22 @@ class LoginScreen(game: BeardBlasterGame) : AbstractScreen(game) {
             this.add(rightTable).width(viewport.worldWidth * 0.9f).fillY()
         }
 
-        // Adding actors to the stage
-        loginStage.addActor(table)
-
+        registrationStage.addActor(table)
     }
 
     override fun update(delta: Float) {
+
     }
 
     override fun setBtnEventListeners() {
-        loginButton.onClick {
-            if (!UserAuth().isLoggedIn() && emailInput.text.isNotEmpty() && passwordInput.text.isNotEmpty()) {
-                UserAuth().signIn(emailInput.text, passwordInput.text)
+        createBtn.onClick {
+            if (emailInput.text.isNotEmpty() && passwordInput.text.isNotEmpty() && userNameInput.text.isNotEmpty()) {
+                UserAuth().createUser(emailInput.text, passwordInput.text, userNameInput.text)
             }
-            // TODO: Future: Don't proceed unless login actually successful
+            // TODO: Future: Don't proceed unless signup actually successful
             game.setScreen<MenuScreen>()
         }
-        backButton.onClick {
+        backBtn.onClick {
             game.setScreen<LoginMenuScreen>()
         }
     }
@@ -145,7 +150,8 @@ class LoginScreen(game: BeardBlasterGame) : AbstractScreen(game) {
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT)
         update(delta)
 
-        loginStage.act(delta)
-        loginStage.draw()
+        registrationStage.act(delta)
+        registrationStage.draw()
+
     }
 }
